Update README.md - Self description
I wrote a self-description in the third person singular. This is part of the activities for my first lesson in GitHub. Although, it is a practice all of the information presented about me are true at the time of writing this. 
Feel free to ask questions about me if you don't understand anything.